Preprocess task: Interview results CSV loading
This is the same conversion process used for the other sections of the
preprocess script, although the changes in this are a bit more involved:

The original script had separated the CSV loading from the code that
transformed/normalized the data into the form used for later exporting
the data back to CSV files.

This was different from how the other sections were written, and was a
bit more convoluted as a result. As part of converting to a Rake task I
simplified things. There is now a single `interview_results` method that
loads the CSV data and transforms it to the proper data structure.