Preprocess task: Early exit on errors
In the original preprocess script we printed any error messages to the
console, but did not stop running the script. As far as I can tell this
is not a good idea because the script is unlikely to work if any of
these errors are encountered.

We now raise a RuntimeError anytime an error happens. Some work could
still be done to convert the complete error output from `puts` to part
of the `raise` message.