Debounce placement saving for 300ms
When changing the placement model (by moving a student between companies
for example) the `onSave` function is called multiple times.

To prevent spamming the server with save requests for intermediate data
states (some of which are invalid), we use Underscore's debounce method
and wait for 300ms after the last call before actually saving.