Spreadsheet.export_data method and test
This method takes an array of headers, an array of arrays of row data,
and a User instance.

It creates a new Google Sheets spreadsheet on the user's Google account
and then uploads the given data into that spreadsheet.

To properly test this you will need to update the User fixture to have a
valid Google OAuth token, at least the first time. After that your tests
should continue to run with the VCR recording.