Hacky fix to support some company names with comma
When a Google Form is exported to a Google Sheet any multi-select
questions have their responses in a single cell with the values
separated by commas.

Our preprocessor script uses this and assumes that commas are valid
delimiters between each of the values in the companies fields.
Unfortunately for C8 a few of the company names include commas and that
is tripping up the preprocessor.

So this commit adds a hacky fix which detects the specific C8 company
names that need to be correct, changes the commas to semicolons, and
then changes them back later after each company name has been split out
properly.