Fix answers data structure in FormResponse tests
We were previously passing in an array of answer objects to the
FormResponse constructor, however that is only the serialized version of
the data.

The constructor is supposed to accept a hash of field_id => answer data
instead, so we needed to create a version like that for the tests.

It's also important that all of the FormResponse.from_params tests still
use the raw/serialized version of the answers structure as they're
dealing with parsing that version into the hash structure.