Restore interviews_per_slot to Classroom
It didn't appear that we were using this anywhere after I replaced the
Classroom loading logic to use the interview schedule CSV instead of
results and preferences spreadsheets.

However, this was because I implemented the loading logic incorrectly in
my new version, setting the slots count to the number of interviews,
when it should be the number of students who can be assigned to that
company.

The classroom loading logic has been fixed to set the correct number of
slots based on the intervews_per_slot value, and tests have been updated
to use appropriate values (otherwise you get companies with slot counts
of 0).