Student rankings don't have to be < 6
We're now allowing students to stack rank their companies, which means
ranking numbers go up to as many interviews as the student had.

We should add tests and validations in the future to ensure no ranking
is higher than the student's interview count, but for now this unbreaks
our tests.