Solver uses Ranking#score
This is part of the work to make that method the canonical source for
the scoring calculation for a particular ranking.

Because the solver needs to use a "cost" factor interpretation of the
score, we first negate the value (making higher scores have a lower
cost).

We also then offset each cost so that the lowest cost (originally the
highest score) has a value of 1, and all other costs increase from
there.